requirements.txt
discord.py~=2.7.1
python-dotenv~=1.2.2
openai~=2.30.0
pydub~=0.25.1
aiohttp~=3.13.5
websockets~=16.0
patreon-v2~=0.5.1
setuptools~=81.0.0